sysinfo: include detected Problems in result directory
This patch adds a new Collectible which collects information from
Problems D-Bus service:
https://jfilak.fedorapeople.org/ProblemsAPI2/

There is currently the only one provider of this D-Bus interface:
http://abrt.readthedocs.io/en/latest/

The result of the new class is either 'core.[pid]' or 'backtrace.[pid]'.
It depends on the type of the problem - e.g. Python problems do not have
core files associated.

The Problems D-Bus service requires authorization via PolicyKit for
allowing access to problems of other users. Hence the current
implementation of ProblemsWatcher gathers only problems of the current
user.

Signed-off-by: Jakub Filak <jfilak@redhat.com>